Move string utils and migrate to typescript
